+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.2.0 - 2022-11-27
+### Added
+- Option to enable target entries on translations publish
+- Option to enable slug copy from source entries to target
+
+### Fixed
+- Multiple drafts apply issue for different sites
+- Draft apply issue (all fields in translation draft now marked as changed for new drafts)
+
 ## 4.1.0 - 2022-10-11
 ### Added
 - Retry logic for failed jobs